Fingers tracking support in VRChat with Valve Index controllers emulation

+++ leap_monitor/Utils.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int ReadEnumVector(const std::string &f_val, const std::vector<std::string> &f_vec);
+size_t ReadEnumVector(const std::string &f_val, const std::vector<std::string> &f_vec);
